+++ test/plugins/elixir/doctest-io-fixtures/clean.ex
@@ -1,0 +1,4 @@
+defmodule Clean do
+  @moduledoc "Just docs."
+  def go, do: :ok
+end