+++ test/plugins/elixir/doctest-io-fixtures/tagged.ex
@@ -1,0 +1,4 @@
+defmodule Tagged do
+  # TODO: For now, this is tracked — should be fine
+  @timeout 5000
+end